Consolidate autosuggest-related modules
