+++ reciperadar/__init__.py
@@ -29,7 +29,6 @@ db = SQLAlchemy(app)
 
 
 import reciperadar.api.feedback
-import reciperadar.api.equipment
-import reciperadar.api.ingredients
+import reciperadar.api.autosuggest
 import reciperadar.api.recipes
 import reciperadar.api.redirect
